+++ source/blender/editors/interface/interface_panel.c
@@ -892,7 +892,7 @@ static void ui_do_animate(const bContext *C, Panel *panel)
 	float fac;
 
 	fac = (PIL_check_seconds_timer() - data->starttime) / ANIMATION_TIME;
-	fac = min_ff(sqrt(fac), 1.0f);
+	fac = min_ff(sqrtf(fac), 1.0f);
 
 	/* for max 1 second, interpolate positions */
 	if (uiAlignPanelStep(sa, ar, fac, false)) {
